feat(consent): Implement ConsentBanner and Tracking components in layout for enhanced user consent management
- Added ConsentBanner component to both play and www apps, encapsulating CookieBanner and ConsentManagerDialog for streamlined consent handling.
- Introduced Tracking component to conditionally render Analytics based on user consent, improving compliance with consent management practices.
- Removed redundant consent-related code from layout.tsx files, promoting cleaner and more maintainable code structure.
